Move experimental tests to relevant folders
