+++ packages/graphql/tests/integration/interfaces/aggegations/aggregation-interfaces-top-level-with-auth.int.test.ts
