+++ packages/graphql/tests/integration/interfaces/aggegations/aggregation-interfaces-top-level.int.test.ts
