+++ packages/graphql/tests/integration/interfaces/relationships/create/create.int.test.ts
@@ -19,8 +19,8 @@
 
 import { faker } from "@faker-js/faker";
 import { generate } from "randomstring";
-import type { UniqueType } from "../../../utils/graphql-types";
-import { TestHelper } from "../../utils/tests-helper";
+import type { UniqueType } from "../../../../utils/graphql-types";
+import { TestHelper } from "../../../utils/tests-helper";
 
 describe("interface relationships", () => {
     const testHelper = new TestHelper();
@@ -29,7 +29,7 @@ describe("interface relationships", () => {
     let Movie: UniqueType;
     let Series: UniqueType;
 
-    beforeEach(async () => {
+    beforeAll(async () => {
         Episode = testHelper.createUniqueType("Episode");
         Actor = testHelper.createUniqueType("Actor");
         Movie = testHelper.createUniqueType("Movie");
@@ -73,11 +73,11 @@ describe("interface relationships", () => {
         });
     });
 
-    afterEach(async () => {
+    afterAll(async () => {
         await testHelper.close();
     });
 
-    test("should disconnect using interface relationship fields", async () => {
+    test("should create create using interface relationship fields", async () => {
         const actorName = generate({
             readable: true,
             charset: "alphabetic",
@@ -90,15 +90,21 @@ describe("interface relationships", () => {
         const movieRuntime = faker.number.int({ max: 100000 });
         const movieScreenTime = faker.number.int({ max: 100000 });
 
-        const seriesTitle = generate({
-            readable: true,
-            charset: "alphabetic",
-        });
-        const seriesScreenTime = faker.number.int({ max: 100000 });
-
         const query = `
-            mutation DisconnectMovie($name: String, $title: String) {
-                ${Actor.operations.update}(where: { name: $name }, disconnect: { actedIn: { where: { node: { title: $title } } } }) {
+            mutation CreateActorConnectMovie($name: String!, $title: String!, $runtime: Int!, $screenTime: Int!) {
+                ${Actor.operations.create}(
+                    input: [
+                        {
+                            name: $name
+                            actedIn: {
+                                create: {
+                                    edge: { screenTime: $screenTime }
+                                    node: { ${Movie}: { title: $title, runtime: $runtime } }
+                                }
+                            }
+                        }
+                    ]
+                ) {
                     ${Actor.plural} {
                         name
                         actedIn {
@@ -112,35 +118,25 @@ describe("interface relationships", () => {
             }
         `;
 
-        await testHelper.executeCypher(
-            `
-                CREATE (a:${Actor} { name: $actorName })
-                CREATE (a)-[:ACTED_IN { screenTime: $movieScreenTime }]->(:${Movie} { title: $movieTitle, runtime:$movieRuntime })
-                CREATE (a)-[:ACTED_IN { screenTime: $seriesScreenTime }]->(:${Series} { title: $seriesTitle })
-            `,
-            {
-                actorName,
-                movieTitle,
-                movieRuntime,
-                movieScreenTime,
-                seriesTitle,
-                seriesScreenTime,
-            }
-        );
-
         const gqlResult = await testHelper.executeGraphQL(query, {
-            variableValues: { name: actorName, title: movieTitle },
+            variableValues: {
+                name: actorName,
+                title: movieTitle,
+                runtime: movieRuntime,
+                screenTime: movieScreenTime,
+            },
         });
 
         expect(gqlResult.errors).toBeFalsy();
 
         expect(gqlResult.data).toEqual({
-            [Actor.operations.update]: {
+            [Actor.operations.create]: {
                 [Actor.plural]: [
                     {
                         actedIn: [
                             {
-                                title: seriesTitle,
+                                runtime: movieRuntime,
+                                title: movieTitle,
                             },
                         ],
                         name: actorName,
@@ -150,12 +146,12 @@ describe("interface relationships", () => {
         });
     });
 
-    test("should nested disconnect using interface relationship fields", async () => {
-        const actorName1 = generate({
+    test("should create create nested nodes using interface relationship fields", async () => {
+        const name1 = generate({
             readable: true,
             charset: "alphabetic",
         });
-        const actorName2 = generate({
+        const name2 = generate({
             readable: true,
             charset: "alphabetic",
         });
@@ -165,24 +161,59 @@ describe("interface relationships", () => {
             charset: "alphabetic",
         });
         const movieRuntime = faker.number.int({ max: 100000 });
-        const movieScreenTime = faker.number.int({ max: 100000 });
+        const screenTime = faker.number.int({ max: 100000 });
 
         const seriesTitle = generate({
             readable: true,
             charset: "alphabetic",
         });
-        const seriesScreenTime = faker.number.int({ max: 100000 });
+
+        const episodeRuntime = faker.number.int({ max: 100000 });
 
         const query = `
-            mutation DisconnectMovie($name1: String, $name2: String, $title: String) {
-                ${Actor.operations.update}(
-                    where: { name: $name1 }
-                    disconnect: {
-                        actedIn: {
-                            where: { node: { title: $title } }
-                            disconnect: { actors: { where: { node: { name: $name2 } } } }
+            mutation CreateActorConnectMovie(
+                $name1: String!
+                $name2: String!
+                $movieTitle: String!
+                $movieRuntime: Int!
+                $screenTime: Int!
+                $seriesTitle: String!
+                $episodeRuntime: Int!
+            ) {
+                ${Actor.operations.create}(
+                    input: [
+                        {
+                            name: $name1
+                            actedIn: {
+                                create: [
+                                    {
+                                        edge: { screenTime: $screenTime }
+                                        node: {
+                                            ${Movie}: {
+                                                title: $movieTitle
+                                                runtime: $movieRuntime
+                                                actors: {
+                                                    create: {
+                                                        edge: { screenTime: $screenTime }
+                                                        node: { name: $name2 }
+                                                    }
+                                                }
+                                            }
+                                        }
+                                    }
+                                    {
+                                        edge: { screenTime: $screenTime }
+                                        node: {
+                                            ${Series}: {
+                                                title: $seriesTitle
+                                                episodes: { create: { node: { runtime: $episodeRuntime } } }
+                                            }
+                                        }
+                                    }
+                                ]
+                            }
                         }
-                    }
+                    ]
                 ) {
                     ${Actor.plural} {
                         name
@@ -194,46 +225,48 @@ describe("interface relationships", () => {
                             ... on ${Movie} {
                                 runtime
                             }
+                            ... on ${Series} {
+                                episodes {
+                                    runtime
+                                }
+                            }
                         }
                     }
                 }
             }
         `;
 
-        await testHelper.executeCypher(
-            `
-                CREATE (a:${Actor} { name: $actorName1 })
-                CREATE (a)-[:ACTED_IN { screenTime: $movieScreenTime }]->(:${Movie} { title: $movieTitle, runtime:$movieRuntime })<-[:ACTED_IN]-(aa:${Actor} { name: $actorName2 })
-                CREATE (a)-[:ACTED_IN { screenTime: $seriesScreenTime }]->(:${Series} { title: $seriesTitle })<-[:ACTED_IN]-(aa)
-            `,
-            {
-                actorName1,
-                actorName2,
+        const gqlResult = await testHelper.executeGraphQL(query, {
+            variableValues: {
+                name1,
+                name2,
                 movieTitle,
                 movieRuntime,
-                movieScreenTime,
+                screenTime,
                 seriesTitle,
-                seriesScreenTime,
-            }
-        );
-
-        const gqlResult = await testHelper.executeGraphQL(query, {
-            variableValues: { name1: actorName1, name2: actorName2, title: movieTitle },
+                episodeRuntime,
+            },
         });
 
         expect(gqlResult.errors).toBeFalsy();
 
         expect(gqlResult.data).toEqual({
-            [Actor.operations.update]: {
+            [Actor.operations.create]: {
                 [Actor.plural]: [
                     {
-                        actedIn: [
+                        actedIn: expect.toIncludeSameMembers([
+                            {
+                                runtime: movieRuntime,
+                                title: movieTitle,
+                                actors: expect.toIncludeSameMembers([{ name: name2 }, { name: name1 }]),
+                            },
                             {
                                 title: seriesTitle,
-                                actors: expect.toIncludeSameMembers([{ name: actorName1 }, { name: actorName2 }]),
+                                actors: [{ name: name1 }],
+                                episodes: [{ runtime: episodeRuntime }],
                             },
-                        ],
-                        name: actorName1,
+                        ]),
+                        name: name1,
                     },
                 ],
             },
