+++ packages/graphql/tests/integration/interfaces/relationships/declare-relationship/interface-two-level-chain.test.ts
@@ -19,8 +19,8 @@
 
 import { faker } from "@faker-js/faker";
 import { gql } from "graphql-tag";
-import type { UniqueType } from "../../../utils/graphql-types";
-import { TestHelper } from "../../utils/tests-helper";
+import type { UniqueType } from "../../../../utils/graphql-types";
+import { TestHelper } from "../../../utils/tests-helper";
 
 describe("interface implementing interface with declared relationships - two level interface chain", () => {
     const testHelper = new TestHelper();
