+++ packages/graphql/tests/integration/interfaces/relationships/declare-relationship/type-narrowing-mutations.test.ts
@@ -19,8 +19,8 @@
 
 import { faker } from "@faker-js/faker";
 import { gql } from "graphql-tag";
-import type { UniqueType } from "../../../utils/graphql-types";
-import { TestHelper } from "../../utils/tests-helper";
+import type { UniqueType } from "../../../../utils/graphql-types";
+import { TestHelper } from "../../../utils/tests-helper";
 
 // TODO: maybe use type-narrowing-connections
 describe("type narrowing - mutations setup", () => {
