+++ packages/graphql/tests/integration/unions/unions.int.test.ts
@@ -20,9 +20,9 @@
 import type { DocumentNode } from "graphql";
 import { gql } from "graphql-tag";
 import { generate } from "randomstring";
-import { createBearerToken } from "../utils/create-bearer-token";
-import type { UniqueType } from "../utils/graphql-types";
-import { TestHelper } from "./utils/tests-helper";
+import { createBearerToken } from "../../utils/create-bearer-token";
+import type { UniqueType } from "../../utils/graphql-types";
+import { TestHelper } from "../utils/tests-helper";
 
 describe("unions", () => {
     const testHelper = new TestHelper();
